Update bike scoring algorithm

+++ src/serializers/addressScoreResult.go
@@ -8,7 +8,7 @@ type AddressScoreResult struct {
 	CBSAName        string  `json:"cbsaName"`
 	NWI             float64 `json:"nwi"`
 	TransitScore    uint8   `json:"transitScore"`
-	BikeScore       uint8   `json:"bikeScore"`
+	BikeScore       float64 `json:"bikeScore"`
 	SearchedAddress string  `json:"searchedAddress"`
 	Format          string  `json:"format"`
 }
@@ -20,7 +20,7 @@ type AddressScoreResultXML struct {
 	CBSAName        string   `xml:"cbsaName"`
 	NWI             float64  `xml:"nwi"`
 	TransitScore    uint8    `xml:"transitScore"`
-	BikeScore       uint8    `xml:"bikeScore"`
+	BikeScore       float64  `xml:"bikeScore"`
 	SearchedAddress string   `xml:"searchedAddress"`
 	Format          string   `xml:"format"`
 }
